Fixed issue, work on adminuser
I used .next instead of .nextLine for the scanner input. This caused spacebars to continue the runtime, so notes and titles could only contain one line.

Furthermore, work on the adminuser was done. It is now possible to list the users in the database, along with their ids and group ids.